Fix smoke test
